Fix formatting of tabs

+++ RaceAnalyser/src/main/java/com/cjsheehan/jrace/racing/Trainer.java
@@ -3,49 +3,48 @@ package com.cjsheehan.jrace.racing;
 import org.apache.commons.lang3.builder.ToStringBuilder;
 
 public class Trainer {
-    private Long id;
-    private String name;
+	private int id;
+	private String name;
 
-    /**
-     * @param name
-     */
-    public Trainer(String name) {
-	super();
-	this.name = name;
-    }
+	/**
+	 * @param name
+	 */
+	public Trainer(String name) {
+		super();
+		this.name = name;
+	}
 
-    /**
-     * @return the id
-     */
-    public Long getId() {
-        return id;
-    }
+	/**
+	 * @return the id
+	 */
+	public int getId() {
+		return id;
+	}
 
-    /**
-     * @param id the id to set
-     */
-    public void setId(Long id) {
-        this.id = id;
-    }
+	/**
+	 * @param id
+	 *            the id to set
+	 */
+	public void setId(int id) {
+		this.id = id;
+	}
 
-    /**
-     * @return the name
-     */
-    public String getName() {
-        return name;
-    }
+	/**
+	 * @return the name
+	 */
+	public String getName() {
+		return name;
+	}
 
-    /**
-     * @param name the name to set
-     */
-    public void setName(String name) {
-        this.name = name;
-    }
-    
-    public String toString() {
-	return new ToStringBuilder(this)
-		.append("Name", this.name)
-		.append("ID", this.id)
-		.toString();
-    }
+	/**
+	 * @param name
+	 *            the name to set
+	 */
+	public void setName(String name) {
+		this.name = name;
+	}
+
+	public String toString() {
+		return new ToStringBuilder(this).append("Name", this.name).append("ID", this.id).toString();
+	}
 }
